Allow implicit conversion in TVM FFI to tvm::Bool (#5907)
